requirements.txt
discord.py
google-genai
python-dotenv
sqlalchemy
psycopg2-binary
apscheduler